requirements.txt
3-1==1.0.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
XlsxWriter==1.3.7
